+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0]
+
 ### Added
 
-- `get_technique`: get a single technique from a name.
+- `get_technique`: get a single technique from a technique identifier.
 
 ### Changed
 
 - `get_ontology_version` and `get_ontology_version_number` can be called without argument.
+- `get_technique_metadata` and `get_techniques` now support both technique names and IRIs as input identifiers.
 
 ### Fixed
 
@@ -26,7 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed ontology base IRI and all entity IRIs from http://purl.org/pan-science/ESRFET to https://w3id.org/PaN/ESRFET.
 - Save technique metadata in HDF5 using the `/NXobject/identifierNAME-field` fields.
 - Versioned IRI stored for the NeXus-compliant format instead of IRI.
-- `get_technique_metadata` and `get_techniques` now support both technique names and IRIs as input identifiers.
 
 ### Added
 
@@ -103,7 +105,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - API for adding scan metadata to HDF5 and dataset metadata to the ESRF data portal.
 - OWL Ontology transpilation for runtime performance.
 
-[unreleased]: https://github.com/pan-ontologies/esrf-ontologies/compare/v2.0.0...HEAD
+[unreleased]: https://github.com/pan-ontologies/esrf-ontologies/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/pan-ontologies/esrf-ontologies/compare/v2.0.0...v2.1.0
 [1.3.0]: https://github.com/pan-ontologies/esrf-ontologies/compare/v1.3.0...v2.0.0
 [1.3.0]: https://github.com/pan-ontologies/esrf-ontologies/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/pan-ontologies/esrf-ontologies/compare/v1.1.0...v1.2.0